Safari-609.1.11.1/LayoutTests/crypto/subtle/rsa-oaep-import-pkcs8-key-expected.txt

Test importing a PKCS8 RSA-OAEP key

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a key...
PASS privateKey.toString() is '[object CryptoKey]'
PASS privateKey.type is 'private'
PASS privateKey.extractable is true
PASS privateKey.algorithm.name is 'RSA-OAEP'
PASS privateKey.algorithm.modulusLength is 2048
PASS bytesToHexString(privateKey.algorithm.publicExponent) is '010001'
PASS privateKey.algorithm.hash.name is 'SHA-1'
PASS privateKey.usages is ['decrypt', 'unwrapKey']
PASS successfullyParsed is true

TEST COMPLETE
